requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
docx==0.2.4
fpdf==1.7.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
lxml==5.1.0
multidict==6.0.5
openai==0.28.0
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.31.0
setuptools==69.1.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
